refs #21 refactor qe functions
